Coded Snake.java class
I used an 800 by 600 pixel canvas for the background, and a circle shape for the snake's head. I used the Pythagorean Theorem to calculate the head's trajectory using the cursor's position and the head's position on the canvas. The game is intended to operate at 60 FPS, and every frame is an update. This means that if your hardware cannot handle the game at 60 FPS, the snake will move slower than intended. The snake's head moves as intended.