updated Snake and added SnakeSegment
In Snake.java, I replaced the "snakeX" and "snakeY" variables with a "SnakeSegment" object, head, which stores its own position and size. The SnakeSegment class was added to keep the snake’s shape and drawing code separate from the main game logic. Now, the Snake class only handles movement and updating, while SnakeSegment handles how the snake is drawn. I mainly updated Snake because I realized I would need a lot more room to add more body parts/segments for the snake, and I can't track the snake's position as effectively if its any bigger than a small circle. As of right now, the game is still effectively the same as it was before updating Snake and adding SnakeSegment