added food, growth, and scoring
added food, food collision mechanics, updated growth mechanics (no longer timed), and added content to the Food class, which is just a little red circle on the board for now. I redid the update module and kept the old timed growth (as comments for reference and future use if necessary). If something happens with food-based growth further down the line I should still be able to use timed growth for testing. The score is displayed on the top left of the screen and is currently updated in increments of 1