added simplified gameplay screens (home, gameplay, pause, game over)
The screens are still basic until I finally add real graphics/effects. Now, when you run the program, you are introduced with a home screen (still a WIP) until you left click to start the game. during the game, you can press space to pause/resume or "X" to end the game. After the game is over, the player's score and time elapsed are displayed, as well as an option to restart.